+++ src/Remote.Linq.Newtonsoft.Json/Properties/AssemblyInfo.cs
@@ -1,5 +1,0 @@
-﻿// Copyright (c) Christof Senn. All rights reserved. See license.txt in the project root for license information.
-
-using System.Security;
-
-[assembly: AllowPartiallyTrustedCallers]